+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.3.0 Under development
 
 - Enh #27: Refactor codebase to improve performance (@terabytesoftw)
+- Bug #28: Add test for setting attributes with prefix value in `HasAttributes` mixin (@terabytesoftw)
 
 ## 0.2.0 March 30, 2024